SoilFactorsAdapter textSize

+++ app/src/main/java/com/example/yandre/biontest/pojo/ElementModel.java
@@ -2,9 +2,9 @@ package com.example.yandre.biontest.pojo;
 
 public class ElementModel {
     private TypeElement element;
-    private int value;
+    private double value;
 
-    public ElementModel(TypeElement element, int value) {
+    public ElementModel(TypeElement element, double value) {
         this.element = element;
         this.value = value;
     }
@@ -17,11 +17,12 @@ public class ElementModel {
         this.element = element;
     }
 
-    public int getValue() {
+
+    public double getValue() {
         return value;
     }
 
-    public void setValue(int value) {
+    public void setValue(double value) {
         this.value = value;
     }
 }
